test_colbert_fix_verification.py
#!/usr/bin/env python3
"""
Verification script to demonstrate that the ColBERT embeddings script fix works correctly.
This script shows the before/after behavior of the token data format handling.
"""

import sys
import os

# Add the project root directory to Python path
project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '.'))
if project_root not in sys.path:
    sys.path.insert(0, project_root)

def test_colbert_encoder_output_format():
    """Test that demonstrates the correct ColBERT encoder output format."""
    print("=== ColBERT Encoder Output Format Verification ===\n")
    
    # Import the ColBERT encoder
    try:
        from archive.colbert.doc_encoder import ColBERTDocEncoder
        print("✅ Successfully imported ColBERTDocEncoder")
        
        # Create a mock encoder
        encoder = ColBERTDocEncoder(mock=True, embedding_dim=128)
        print("✅ Successfully created mock ColBERT encoder")
        
        # Test encoding
        test_text = "This is a test document for ColBERT token embedding generation."
        result = encoder.encode(test_text)
        
        print(f"\n📝 Test text: '{test_text}'")
        print(f"🔍 Encoder output type: {type(result)}")
        print(f"🔍 Encoder output structure: {type(result)} with {len(result)} elements")
        
        if isinstance(result, tuple) and len(result) == 2:
            tokens, embeddings = result
            print(f"✅ Correct tuple format: (tokens, embeddings)")
            print(f"   - Tokens type: {type(tokens)}, count: {len(tokens)}")
            print(f"   - Embeddings type: {type(embeddings)}, count: {len(embeddings)}")
            print(f"   - First few tokens: {tokens[:3]}")
            print(f"   - First embedding shape: {len(embeddings[0]) if embeddings else 0} dimensions")
            
            # Show how the script now processes this
            print(f"\n🔧 Script processing:")
            print(f"   - Receives tuple: {type(result)}")
            print(f"   - Extracts tokens and embeddings separately")
            print(f"   - Creates token_data = list(zip(tokens, embeddings))")
            
            token_data = list(zip(tokens, embeddings))
            print(f"   - Result: {len(token_data)} (token, embedding) pairs")
            print(f"   - First pair: ('{token_data[0][0]}', embedding[{len(token_data[0][1])}])")
            
            return True
        else:
            print(f"❌ Unexpected format: {result}")
            return False
            
    except Exception as e:
        print(f"❌ Error: {e}")
        return False

def test_script_function():
    """Test the actual script function with mock data."""
    print("\n=== Script Function Verification ===\n")
    
    try:
        from scripts.populate_missing_colbert_embeddings import process_single_document
        from unittest.mock import MagicMock
        
        print("✅ Successfully imported script functions")
        
        # Create mock database connection
        mock_conn = MagicMock()
        mock_cursor = MagicMock()
        mock_conn.cursor.return_value = mock_cursor
        
        # Create mock ColBERT encoder that returns the correct format
        mock_encoder = MagicMock()
        mock_encoder.return_value = (
            ["token1", "token2", "token3"],  # tokens
            [[0.1, 0.2, 0.3], [0.4, 0.5, 0.6], [0.7, 0.8, 0.9]]  # embeddings
        )
        
        # Test document
        test_doc = {
            'doc_id': 'TEST_DOC_001',
            'text_content': 'This is a test document for verification.',
            'abstract': None,
            'title': None
        }
        
        print(f"📝 Test document: {test_doc['doc_id']}")
        print(f"🔧 Mock encoder returns: (tokens_list, embeddings_list)")
        
        # Call the function
        result = process_single_document(test_doc, mock_conn, mock_encoder)
        
        print(f"✅ Function executed successfully: {result}")
        print(f"✅ Mock encoder was called: {mock_encoder.called}")
        print(f"✅ Database operations were attempted: {mock_cursor.executemany.called}")
        
        return True
        
    except Exception as e:
        print(f"❌ Error: {e}")
        import traceback
        traceback.print_exc()
        return False

def main():
    """Run all verification tests."""
    print("ColBERT Token Embedding Script Fix Verification")
    print("=" * 50)
    
    test1_passed = test_colbert_encoder_output_format()
    test2_passed = test_script_function()
    
    print(f"\n{'=' * 50}")
    print("VERIFICATION SUMMARY:")
    print(f"✅ ColBERT Encoder Format Test: {'PASSED' if test1_passed else 'FAILED'}")
    print(f"✅ Script Function Test: {'PASSED' if test2_passed else 'FAILED'}")
    
    if test1_passed and test2_passed:
        print(f"\n🎉 ALL TESTS PASSED! The fix is working correctly.")
        print(f"\nThe script now correctly handles the ColBERT encoder output format:")
        print(f"  - Encoder returns: (List[str], List[List[float]])")
        print(f"  - Script converts to: List[Tuple[str, List[float]]]")
        print(f"  - Database storage works as expected")
    else:
        print(f"\n❌ Some tests failed. Please check the implementation.")
    
    return test1_passed and test2_passed

if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)