requirements.txt
# Core IRIS and ML dependencies
intersystems-irispython==5.1.2
sqlalchemy>=2.0.0
testcontainers-iris>=1.2.0
testcontainers>=3.7.0

# Machine Learning and Embeddings
sentence-transformers>=2.3,<3
huggingface-hub>=0.20,<1
torch>=2.7.0
torchvision>=0.22.0
torchaudio>=2.7.0
transformers>=4.52.3
accelerate>=1.7.0

# LLM Integration
langchain>=0.1.0
langchain-openai>=0.1.0,<0.2.0

# Data Processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0

# Evaluation and Monitoring
ragas>=0.1.0
evidently>=0.4.0

# Graph Processing
networkx>=3.1

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
psutil>=5.9.0
PyYAML>=6.0
jaydebeapi>=1.2.3
pyodbc>=5.2.0

# Visualization
matplotlib>=3.10.3
seaborn>=0.13.0
plotly>=5.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0
pytest-lazy-fixture>=0.6.3
pytest-mock>=3.14.0
pytest-asyncio>=0.21.0

# Development Tools
ruff>=0.1.5
black>=23.11.0
mypy>=1.7.0

# Optional AWS Support
boto3>=1.28.0